Build in a docker container with Debian Stretch, not Jessie
This is necessary for the resulting packages to work on Stretch
(which is the current Debian stable release) due to the different
version of the libMagickWand library.

Later, we will want to target other distros and distro versions too.
We'll either have to make the CI pipeline more complex (build each
package in its appropriate Docker container) or link libmagickwand
statically.

For now it's ok if we have a working package for Debian stable.