Auto-orient thumbnails, always serve thumbnails
Orientation handling is difficult to get right in the browser due to
different orientation handling in different browsers and the fact that
neither JustifiedGallery nor PhotoSwipe have any support for this.

The first approach I cam up with was using CSS transforms to rotate
the gallery thumbnails (based on information from the backend).
But then we'd need to tell PhotoSwipe to use width/height for the
layout calculation and height/width (swapped) for the actual image
size. This is not possible yet. The approach seems like much more work.

Instead, we now auto-orient thumbnails on the server side.  This is
already enough for gallery. To make it work in the slideshow too, we
force PhotoSwipe to always show thumbnails.

The side benefit of that is that very large pictures don't slow the
slideshow down anymore. Additionally, later we will be able to easily
add optimizations for mobile and high-dpi screens.